pyproject.toml
[tool.poetry]
name = "iamages_server"
version = "4.0.0"
description = "Server for the Iamages service."
authors = ["Nam Thành Nguyễn <35841468+jkelol111@users.noreply.github.com>"]
license = "GPLv3"

[tool.poetry.dependencies]
python = "^3.10"
fastapi = "^0.79.0"
passlib = {extras = ["argon2", "bcrypt"], version = "^1.7.4"}
python-multipart = "^0.0.5"
Jinja2 = "^3.1.2"
python-magic = "^0.4.27"
motor = "^3.0.0"
brotli-asgi = "^1.1.0"
opencv-python-headless = "^4.6.0"
exif = "^1.3.5"
orjson = "^3.7.11"
python-jose = {extras = ["pycryptodome"], version = "^3.3.0"}
uvicorn = {extras = ["standard"], version = "^0.18.2"}

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
